Use new support for XML part 1
git-svn-id: http://goldengate.svn.sourceforge.net/svnroot/goldengate/GoldenGateR66/trunk@662 7864f6d1-e753-4b29-85b9-74ae93846b71

+++ src/main/java/openr66/database/data/DbRule.java
@@ -842,7 +842,7 @@ public class DbRule extends AbstractDbData {
             nodedelay = noderoot.selectSingleNode(TASK_DELAY);
             String delay;
             if (nodedelay == null) {
-                delay = Integer
+                delay = Long
                         .toString(Configuration.configuration.TIMEOUTCON);
             } else {
                 delay = nodedelay.getText();
